Fix a kernel warning under FreeBSD.
This fixes messages like below in dmesg when running cfs(1):

WARNING pid 36592 (cfs): ioctl sign-extension ioctl ffffffff80245604

+++ coda-src/kerndep/pioctl.c
@@ -257,7 +257,7 @@ int pioctl(const char *path, unsigned long com,
     /* Must change the size field of the command to match 
        that of the new structure. */
     unsigned long cmd = (com & ~(PIOCPARM_MASK << 16)); /* mask out size  */
-    int	size = ((com >> 16) & PIOCPARM_MASK) + sizeof(char *) + sizeof(int);
+    unsigned long size = ((com >> 16) & PIOCPARM_MASK) + sizeof(char *) + sizeof(int);
 
     cmd	|= (size & PIOCPARM_MASK) << 16;  /* or in corrected size */
